# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
        "/api/v1/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register new user",
                "operationId": "register",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "name of the new user",
                                        "type": "string"
                                    },
                                    "email": {
                                        "description": "email of the new user",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "password of the new user",
                                        "type": "string"
                                    },
                                    "password_confirmation": {
                                        "description": "confirmation of user's password",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User registered successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "title": "data",
                                            "description": "Response data",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "access_token": {
                                                        "title": "access token",
                                                        "description": "Access token",
                                                        "type": "string"
                                                    },
                                                    "refresh_token": {
                                                        "title": "refresh token",
                                                        "description": "Refresh token",
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "title": "message",
                                            "description": "Response message",
                                            "type": "string"
                                        },
                                        "status": {
                                            "title": "status",
                                            "description": "Response status",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Validation error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "title": "data",
                                            "description": "Response data",
                                            "type": "array",
                                            "items": {}
                                        },
                                        "message": {
                                            "title": "message",
                                            "description": "Response message",
                                            "type": "string"
                                        },
                                        "status": {
                                            "title": "status",
                                            "description": "Response status",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user",
                "operationId": "login",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "description": "email of the new user",
                                        "type": "string"
                                    },
                                    "password": {
                                        "description": "password of the new user",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User logged in successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "title": "data",
                                            "description": "Response data",
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "access_token": {
                                                        "title": "access token",
                                                        "description": "Access token",
                                                        "type": "string"
                                                    },
                                                    "refresh_token": {
                                                        "title": "refresh token",
                                                        "description": "Refresh token",
                                                        "type": "string"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "message": {
                                            "title": "message",
                                            "description": "Response message",
                                            "type": "string"
                                        },
                                        "status": {
                                            "title": "status",
                                            "description": "Response status",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unathorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "title": "data",
                                            "description": "Response data",
                                            "type": "array",
                                            "items": {}
                                        },
                                        "message": {
                                            "title": "message",
                                            "description": "Response message",
                                            "type": "string"
                                        },
                                        "status": {
                                            "title": "status",
                                            "description": "Response status",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/logout": {
            "delete": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout the current user",
                "operationId": "logout",
                "requestBody": {
                    "description": "Pass user credentials",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "token"
                                ],
                                "properties": {
                                    "token": {
                                        "description": "Access token of the user",
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User successfully logged out",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "title": "data",
                                            "description": "Response data",
                                            "type": "array",
                                            "items": {}
                                        },
                                        "message": {
                                            "title": "message",
                                            "description": "Response message",
                                            "type": "string"
                                        },
                                        "status": {
                                            "title": "status",
                                            "description": "Response status",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unathorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "title": "data",
                                            "description": "Response data",
                                            "type": "array",
                                            "items": {}
                                        },
                                        "message": {
                                            "title": "message",
                                            "description": "Response message",
                                            "type": "string"
                                        },
                                        "status": {
                                            "title": "status",
                                            "description": "Response status",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "login/{driver}": {
            "get": {
                "tags": [
                    "Third-party authorization"
                ],
                "summary": "User autorization with social networks",
                "operationId": "loginThirdParty",
                "responses": {
                    "200": {
                        "description": "User's changed personal info"
                    },
                    "401": {
                        "description": "Unathorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "title": "data",
                                            "description": "Response data",
                                            "type": "array",
                                            "items": {}
                                        },
                                        "message": {
                                            "title": "message",
                                            "description": "Response message",
                                            "type": "string"
                                        },
                                        "status": {
                                            "title": "status",
                                            "description": "Response status",
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "passport": {
                "type": "oauth2",
                "description": "Laravel passport oauth2 security.",
                "in": "header",
                "scheme": "https",
                "flows": {
                    "password": {
                        "authorizationUrl": "http://app.test/oauth/authorize",
                        "tokenUrl": "http://app.test/oauth/token",
                        "refreshUrl": "http://app.test/token/refresh",
                        "scopes": []
                    }
                }
            }
        }
    },
    "security": [
        []
    ]
}